# server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_hybrid-app_server_internal_httpx.Envelope:
    properties:
      code:
        description: 0 = 成功，非 0 = 业务错误码（沿用 HTTP 状态码语义）
        type: integer
      data:
        description: 业务数据
      message:
        description: 人类可读信息
        type: string
    type: object
...
  internal_handler.exportDevicesByIDsReq:
    properties:
      format:
        type: string
      ids:
        items:
          type: integer
        type: array
    type: object
...
  internal_handler.registerDeviceReq:
    properties:
      adid:
        type: string
      appId:
        type: string
      deviceKey:
        type: string
      deviceName:
        type: string
      gaid:
        type: string
      oaid:
        type: string
      palcode:
        type: string
    type: object
...
info:
  contact: {}
  description: hybrid-app 多渠道打包后台：渠道 CRUD / 图标管线 / 域名配置 / 运行时下发。
  title: 渠道中台 API
  version: "1.0"
paths:
...
  /api/app/device/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 设备上报请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.registerDeviceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      summary: APK 上报设备信息（公开端点）
      tags:
      - device
...
  /api/channels/{id}/icon:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 主图 PNG/JPEG
        in: formData
        name: file
        required: true
        type: file
      - description: '自适应图标背景色 #RRGGBB'
        in: formData
        name: background
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上传裁剪后的主图(1024²) → imaging 生成全密度 → 返回各槽位预览
      tags:
      - icon
...
  /api/channels/{id}/res.zip:
    get:
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 下载整套 res 资源 zip（CLI/构建机用；越界返回 404，runner 不受数据权限约束）
      tags:
      - icon
  /api/channels/{id}/splash:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: splash 源图
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上传 splash 源图
      tags:
      - icon
  /api/devices:
    get:
      parameters:
      - description: 按 applicationId 精确筛选
        in: query
        name: applicationId
        type: string
      - description: 起始日期 YYYY-MM-DD（含）
        in: query
        name: from
        type: string
      - description: 结束日期 YYYY-MM-DD（含）
        in: query
        name: to
        type: string
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认 50，最大 200
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 设备列表（分页/筛选，page:devices）
      tags:
      - device
  /api/devices/export:
    post:
      consumes:
      - application/json
      parameters:
      - description: id 列表（最多 1000 个）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.exportDevicesByIDsReq'
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 按勾选 id 导出设备 CSV（device:export，同构流式响应）
      tags:
      - device
  /api/devices/export-token:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 签发设备 CSV 导出下载令牌（device:export，5 分钟有效）
      tags:
      - device
  /api/devices/export.csv:
    get:
      parameters:
      - description: GET /api/devices/export-token 换来的令牌
        in: query
        name: token
        required: true
        type: string
      - description: 按 applicationId 精确筛选
        in: query
        name: applicationId
        type: string
      - description: 起始日期 YYYY-MM-DD（含）
        in: query
        name: from
        type: string
      - description: 结束日期 YYYY-MM-DD（含）
        in: query
        name: to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 设备 CSV 导出（按筛选条件流式下载；用 export-token 换来的一次性令牌鉴权，不接受 access token）
      tags:
      - device
  /api/devices/export.xlsx:
    get:
      parameters:
      - description: GET /api/devices/export-token 换来的令牌
        in: query
        name: token
        required: true
        type: string
      - description: 按 applicationId 精确筛选
        in: query
        name: applicationId
        type: string
      - description: 起始日期 YYYY-MM-DD（含）
        in: query
        name: from
        type: string
      - description: 结束日期 YYYY-MM-DD（含）
        in: query
        name: to
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 设备 XLSX 导出（美化表头/列宽/冻结首行；按筛选条件下载，鉴权同 export.csv）
      tags:
      - device
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"